+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Added
 
 - **Chat Frame:** Option to hide the messages of "learned" and "unlearned" spells on respec
+- **Instance Difficulty:** Change the location, color and size of the dungeon difficulty text
 
 ---